chore: remove the migration scripts that can no longer run
The parser, classifier, link rewriter, renderer, manifest gate and issue
writer all read the wiki checklists, and Phase 4 of the migration deleted
those checklists. They cannot be run again, so leaving them in the tree
only invites someone to try -- three of main.ts's commands already exited
0 while doing nothing.

What stays is what still works: labels.sh recreates the label set,
backfill-project.ts puts issues on the board until auto-add is enabled,
and Throttle moves out of github.ts into its own module because the
backfill is its last caller. The four migrate:* package scripts go with
main.ts. The runbook records which three files are left and why, and
.migration/created.json stays tracked -- it is the only record of which
checklist line became which issue.
